ProjMgtService-Refactoring: Refactored package names for DTOs.
