+++ src/main/java/com/umssonline/proymgt/models/dto/backlog/UpdateBacklogDto.java
@@ -1,4 +1,4 @@
-package com.umssonline.proymgt.models.dto;
+package com.umssonline.proymgt.models.dto.backlog;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import io.swagger.annotations.ApiModel;
