+++ src/main/java/com/umssonline/proymgt/models/dto/project/UpdateProjectDto.java
@@ -1,7 +1,7 @@
-package com.umssonline.proymgt.models.dto;
+package com.umssonline.proymgt.models.dto.project;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.umssonline.proymgt.models.dto.backlog.CreateBacklogDto;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Getter;
@@ -11,12 +11,19 @@ import javax.validation.constraints.Future;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
-@ApiModel(description = "Project Model")
+@ApiModel(description = "Backlog")
 @Getter
 @Setter
-public class CreateProjectDto {
+public class UpdateProjectDto {
+
+    @ApiModelProperty
+    (
+        notes = "Id field should not be null.",
+        required = true
+    )
+    @NotNull(message = "Id field should not be null.")
+    private Long Id;
 
     @ApiModelProperty
     (
@@ -28,10 +35,10 @@ public class CreateProjectDto {
     private String name;
 
     @ApiModelProperty
-            (
-                    notes = "Completed Date Estimation field can should not be null, and It should be future date.",
-                    required = true
-            )
+    (
+        notes = "Completed Date Estimation field can should not be null, and It should be future date.",
+        required = true
+    )
     @NotNull(message = "Completed Date Estimation field can should not be null.")
     @Future(message = "Completed Date Estimation should be a future date.")
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy")
@@ -45,7 +52,4 @@ public class CreateProjectDto {
     @NotNull(message = "Backlog should not be null.")
     private CreateBacklogDto backlog;
 
-
-    @JsonIgnore
-    private LocalDateTime createdAt = LocalDateTime.now();
 }
